requirements.txt
git+https://github.com/jvsl/django-audit-log.git
Django==1.8.5
Markdown==2.6.2
MySQL-python==1.2.5
PyJWT==1.4.0
PyYAML==3.11
argparse==1.2.1
astroid==1.3.8
boto==2.30.0
coverage==3.7.1
django-boto==0.3.1
django-cors-headers==1.1.0
django-db-signals==0.1.1
django-extensions==1.5.7
django-extra-fields==0.4
django-filter==0.11.0
django-jenkins==0.17.0
django-nose==1.4.2
django-storages==1.1.8
djangorestframework==3.2.3
djangorestframework-filters==0.5.0
djangorestframework-jwt==1.7.2
funcsigs==0.4
incf.countryutils==1.0
logilab-common==1.0.2
mock==1.0.1
nose==1.3.7
pbr==1.8.1
pep8==1.6.2
pycountry==1.20
pyflakes==0.9.2
pylint==1.4.4
python-dateutil==2.4.2
python-gcm==0.4
pytz==2015.6
six==1.9.0
sqlparse
uWSGI==2.0.11.2
tzlocal==1.2.2
wheel==0.24.0
wsgiref==0.1.2


